reorganized code for alteration operation and biomarker prediction
1: In the previous version, when user need to add a new alteration, it
will send a alteration existence check from front end. based on the
result, it will decide to insert alteration record or not and then will
send mapping record existence check from front end. Things like this
going back and forth from front end to the back end is unnecessary
since the curator didn’t care out the detailed procedure and he just
need to know the result, which is if the alteration record get added
successfully. The code structure before was not only not efficient but
also hard to organize. Thus I decided to to rewrite the add and delete
alteration function. The new structure moved most of the work to the
back end and keep front and back end clear.

2: for the biomarker prediction, the previous version divided genes
into two groups, genes having mutation record and not having any
mutation records. But this model has an obvious loophole, like for a
gene having mutation record like BRAF V600E, a trial might only match
that gene. In that case, the record won’t be detected. At first I just
added a new condition check for the case but then I realize that model
doesn’t make sense from design. So I just rewrite a new model that will
check gene one by one. The details about this model can be found on the
google doc, and its link can be found in the googleDoc.txt